refactor: VoiceState.ts -> VoiceState.svelte.ts
