+++ src/lib/components/VideoUI.svelte
@@ -1,20 +1,20 @@
 <script lang="ts">
-  import { voiceState } from "$lib/voice/VoiceState";
+  import { voiceState } from "$lib/voice/VoiceState.svelte";
   import MicAction from "./actions/MicAction.svelte";
-  import Avatar from "./Avatar.svelte";
   import VideoTrack from "./VideoTrack.svelte";
   import { User } from "uprising.js";
   import VideoAction from "./actions/VideoAction.svelte";
   import ScreencastAction from "./actions/ScreencastAction.svelte";
   import UserDisplay from "./UserDisplay.svelte";
+  import { getContext } from "svelte";
 
   //let stream = voiceState.stream;
   //let streams = voiceState.streams;
   let participants = voiceState.participants;
   $inspect(participants);
+  let me = getContext<User>("user");
   let users = $derived.by(() =>{
     const keys = [...participants.keys()];
-    console.debug("Participants (ids)", keys);
     return new Map(
       keys.map((id) => [
         id,
@@ -29,7 +29,7 @@
     <div class="row">
       {#if data.streams.length}
         {#each data.streams as stream (stream.id)}
-            <VideoTrack stream={stream} user={users.get(id)} voiceUser={data} />
+            <VideoTrack stream={stream} user={users.get(id)} voiceUser={data} muted={me.id == id} />
         {/each}
       {:else}
         <UserDisplay user={users.get(id)} />
