+++ src/lib/components/actions/EndCallAction.svelte
@@ -1,12 +1,12 @@
 <script>
-    import { voiceState } from "$lib/voice/VoiceState";
+    import { voiceState } from "$lib/voice/VoiceState.svelte";
     import Icon from "@iconify/svelte";
     import { getContext } from "svelte";
     import { IconButton } from "vermeer-ui";
 
     const user = getContext("user");
     function leaveCall() {
-        voiceState.leave(user?.id);
+        voiceState.leave();
     }
 </script>
